Refector) App.jsx : 관심사 분리를 위한 Input 컴포넌트 및 엘리먼트 이동
